Moved input to a function, moved some containers to a file
