+++ src/Input.hpp
@@ -1,0 +1,10 @@
+#ifndef INPUT_H
+#define INPUT_H
+
+#include <glm/glm.hpp>
+#include <SDL2/SDL.h>
+#include "Containers.hpp"
+
+void handle_keyboard_input(Camera &camera, bool &loop);
+
+#endif
