+++ src/Model.hpp
@@ -1,5 +1,5 @@
-#ifndef MODEL
-#define MODEL
+#ifndef MODEL_H
+#define MODEL_H
 
 #include <SDL2/SDL_image.h>
 #include "Utils.hpp"
