+++ src/main.cpp
@@ -1,5 +1,7 @@
 #include "Utils.hpp"
 #include "Model.hpp"
+#include "Input.hpp"
+#include "Containers.hpp"
 
 GLuint shader_forward, shader_geometry, shader_deferred;
 GLuint screen_width = 800;
@@ -7,21 +9,19 @@ GLuint screen_height = 640;
 SDL_Window* main_window;
 SDL_GLContext main_context;
 
-glm::vec3 camera_pos = glm::vec3(0.f, 0.f, 3.f);
-glm::vec3 camera_front = glm::vec3(0.f, 0.f, -1.f);
-glm::vec3 camera_up = glm::vec3(0.0f, 1.0f, 0.0f);
-glm::mat4 camera_rot;
-
-float camera_speed = 0.1f;
-float camera_rot_speed = 0.01f;
+Camera camera = {glm::vec3(0.f, 0.f, 3.f),    // Position
+                 glm::vec3(0.f, 0.f, -1.f),   // Front
+                 glm::vec3(0.0f, 1.0f, 0.0f), // Up
+                 glm::vec3(1.f, 0.f, 0.f),    // Right
+                 0.3f,                        // Speed
+                 0.01f,                       // Rotational speed
+};
 
 glm::mat4 w2v_matrix;
 glm::mat4 projection_matrix;
 
-
 std::vector<Model*> loaded_models;
 std::vector<Light*> loaded_lights;
-
 
 void run_game();
 
@@ -42,7 +42,7 @@ int main(int argc, char *argv[])
     upload_lights(shader_forward, loaded_lights);
 
     // Create and upload view uniforms:
-    w2v_matrix = glm::lookAt(camera_pos, camera_pos + camera_front, camera_up);
+    w2v_matrix = glm::lookAt(camera.position, camera.position + camera.front, camera.up);
     projection_matrix = glm::perspective(45.0f, (float)screen_width / (float)screen_height, 0.1f, 100.0f);
     GLint view = glGetUniformLocation(shader_forward, "view");
     glUniformMatrix4fv(view, 1, GL_FALSE, glm::value_ptr(w2v_matrix));
@@ -61,70 +61,9 @@ void run_game() {
     bool loop = true;
 
     while (loop) {
-        SDL_Event event;
-        const Uint8* keystate = SDL_GetKeyboardState(NULL);
-        glm::vec3 camera_right = glm::normalize(glm::cross(camera_front, camera_up));
+        handle_keyboard_input(camera, loop);
 
-        if(keystate[SDL_GetScancodeFromKey(SDLK_w)]) {
-            camera_pos += camera_speed * camera_front;
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_s)]) {
-            camera_pos -= camera_speed * camera_front;
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_a)]) {
-            camera_pos -= glm::normalize(glm::cross(camera_front, camera_up)) * camera_speed;
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_d)]) {
-            camera_pos += glm::normalize(glm::cross(camera_front, camera_up)) * camera_speed;
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_q)]) {
-            camera_rot = glm::rotate(glm::mat4(1.0f),
-                                     camera_rot_speed,
-                                     glm::vec3(0.f, 1.f, 0.f) //camera_up
-                                     );
-            camera_front = glm::vec3(camera_rot * glm::vec4(camera_front, 1.f));
-            camera_up = glm::vec3(camera_rot * glm::vec4(camera_up, 1.f));
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_e)]) {
-            camera_rot = glm::rotate(glm::mat4(1.0f),
-                                     -camera_rot_speed,
-                                     glm::vec3(0.f, 1.f, 0.f) //camera_up
-                                     );
-            camera_front = glm::vec3(camera_rot * glm::vec4(camera_front, 1.f));
-            camera_up = glm::vec3(camera_rot * glm::vec4(camera_up, 1.f));
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_f)]) {
-            camera_right = glm::normalize(glm::cross(camera_front, camera_up));
-            camera_rot = glm::rotate(glm::mat4(1.0f),
-                                     -camera_rot_speed,
-                                     camera_right
-                                 );
-            camera_front = glm::vec3(camera_rot * glm::vec4(camera_front, 1.f));
-            camera_up = glm::vec3(camera_rot * glm::vec4(camera_up, 1.f));
-        }
-        if(keystate[SDL_GetScancodeFromKey(SDLK_r)]) {
-            camera_right = glm::normalize(glm::cross(camera_front, camera_up));
-            camera_rot = glm::rotate(glm::mat4(1.0f),
-                                     camera_rot_speed,
-                                     camera_right
-                                 );
-            camera_front = glm::vec3(camera_rot * glm::vec4(camera_front, 1.f));
-            camera_up = glm::vec3(camera_rot * glm::vec4(camera_up, 1.f));
-        }
-
-        while (SDL_PollEvent(&event)) {
-            if (event.type == SDL_QUIT)
-                loop = false;
-            if (event.type == SDL_KEYDOWN) {
-                switch (event.key.keysym.sym) {
-                case SDLK_ESCAPE:
-                    loop = false;
-                    break;
-                }
-            }
-        }
-
-        w2v_matrix = glm::lookAt(camera_pos, camera_pos + camera_front, camera_up);
+        w2v_matrix = glm::lookAt(camera.position, camera.position + camera.front, camera.up);
         GLint view = glGetUniformLocation(shader_forward, "view");
         glUniformMatrix4fv(view, 1, GL_FALSE, glm::value_ptr(w2v_matrix));
